Conflicts:
	src/game.cpp
	src/player.cpp

+++ src/player_activity.cpp
@@ -77,9 +77,9 @@ bool player_activity::is_suspendable() const {
 
 
 int player_activity::get_value(int index, int def) const {
-    return (index < values.size()) ? values[index] : def;
+    return (index < (ssize_t)values.size()) ? values[index] : def;
 }
 
 std::string player_activity::get_str_value(int index, std::string def) const {
-    return (index < str_values.size()) ? str_values[index] : def;
+    return (index < (ssize_t)str_values.size()) ? str_values[index] : def;
 }
